big, big, big time refactoring
Restructured entire project to be way more sensible/extensible. Reworked a lot of core behaviors of modules so that they can be running directly from command line, reworked a lot of default behaviors to be more sane.

Still more work to do in terms of default output filenames but, overall, coming together nicely.